+++ fractal/components/tooltips/ToolTip/ToolTip.config.js
@@ -4,8 +4,6 @@ module.exports = {
     "collated": true,
     "preview": "@tool-tip-preview",
     "context": {
-        "className": "push-left",
-        "label": "Tool Tip Title",
         "text": "Lorem ipsum dolor sit amet, consectetur adipiscing elit.",
         "expanded": true
     }