+++ fractal/components/tooltips/ToolTip/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import ToolTip from './ToolTip.jsx';
+import config from './ToolTip.config';
+
+ReactDOM.render(
+    <ToolTip {...config.context}/>,
+    document.getElementsByClassName('base-preview')[0]
+);